28/09/2019 -  Car and Resolution Update
• Removed the barrels;
• Removed barrel.png;
• Added otherCars.png, along with four variations;
• Added cars on the road to replace the barrels, each one having a different speed;
• Added resolution.cpp and resolution.h;
• Resolution and fullscreen settings now save on quitting the game;
• Added initImg() and destroy();
• The player's hitbox now gets smaller when performing curves;
• The entire barrel mechanic has been adapted to fit the NPCCar array.

+++ src/text.cpp
@@ -13,10 +13,10 @@ void writeText(SDL_Renderer *render, int text, SDL_Texture *font, int x, int y){
 }
 
 void writeText(SDL_Renderer *render, std::string text, SDL_Texture *font, int x, int y, int w = 28, int h = 48){
-    int i = 0;
+    unsigned long long i = 0;
     while(text[i] != 0){
         SDL_Rect textCut = {(text[i] - 48) * 16, 0, 16, 24};
-        SDL_Rect textPos = {x + ((w-2) * i), y, w, h};
+        SDL_Rect textPos = {x + ((w-2) * static_cast<int>(i)), y, w, h};
         SDL_RenderCopy(render, font, &textCut, &textPos);
         i++;
     }
